requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-cors-headers==2.2.0
django-sslserver==0.20
djangorestframework==3.8.2
ecdsa==0.13
future==0.16.0
idna==2.6
opentok==2.6.0
Pillow==5.1.0
psycopg2==2.7.4
pyasn1==0.4.2
python-jose==3.0.0
python-memcached==1.59
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22

